Create a new component for the search modifiers
